requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-APScheduler==1.12.4
web3==6.5.0
eth-account==0.9.0
qrcode==7.4.2
Pillow==10.0.0
python-dotenv==1.0.0
